% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_and_mask_fuel_raster.R
\name{align_and_mask_fuel_raster}
\alias{align_and_mask_fuel_raster}
\title{Align and mask a fuel biomass raster to lidar-derived processing mask}
\usage{
align_and_mask_fuel_raster(fuel_rast, voxel_mask, method = "bilinear")
}
\arguments{
\item{fuel_rast}{A \code{SpatRaster} containing fuel biomass values.
Must have a valid coordinate reference system.}

\item{voxel_mask}{A \code{SpatRaster} produced by
\verb{\link{build_voxel_mask}}, defining lidar-eligible areas
(values of 1) and ineligible areas (\code{NA}).}

\item{method}{Character string specifying the resampling method used
if reprojection is required. Passed to \code{terra::project()}.
Defaults to \code{"bilinear"}.}
}
\value{
A \code{SpatRaster} representing the fuel biomass raster
reprojected (if necessary), cropped to the lidar extent, and masked
to eligible voxels.
}
\description{
This function reprojects, crops, and masks a 2D fuel biomass raster so that
it is spatially aligned with a lidar-derived voxel processing mask.
The output raster matches the lidar extent, resolution, and CRS, with
ineligible areas set to \code{NA}.
}
\details{
If the coordinate reference system (CRS) of \code{fuel_rast} does not
match that of \code{voxel_mask}, the fuel raster is reprojected to
the voxel mask CRS prior to cropping and masking.
}
\examples{
\dontrun{
library(terra)
fuel_file <- system.file("extdata", "fb_sample.tif", package = "voxelFuel")
las_file <- system.file("extdata", "lidar_sample.laz", package = "voxelFuel")
fuel_rast <- rast(fuel_file)
las <- readLAS(las_file)
voxel_mask <- build_voxel_mask(las, res = 10, min_density = 10, min_height = 5)
aligned_fuel <- align_and_mask_fuel_raster(fuel_rast, voxel_mask)
}

}
